New spell: Kindle Blastsparks (L3 Fire/Air/Tloc)
Conjure Flame is a strong effect, but the current incarnation has
proven confusing and unintuitive for players. Let's try something a
little bit different.

Kindle Blastsparks creates a cloud of blastsparks on the caster's tile.
When any creature moves *into* that cloud (eg a monster chasing the
caster), or when that cloud is hit by fire damage (fireball, wand of
flame, etc), it explodes in a 3x3 blast. The blast does low damage, but
it also knocks back any creature not at its center, pushing them away.
This allows for a whole bunch of cute tricks.

Damage currently scales with power from 2d10 to 2d18ish. If players often
find it undesirable to increase power, because they'd rather blast
themselves as a cheeky movement trick, then we can instead scale cloud dur
or knockback chance on power.

Replaces Conjure Flame in the FE start book. It's a bit unsynergistic
with HW's spells (imp, meph...), so I want to do something else there.