New spell: Cigotuvi's Dreadful Rot (L2 Nec/Air/Pois)
Part B of the plan to replace Conjure Flame and Corpse Rot, Cigotuvi's
Dreadful Rot plops a single-tile cloud of miasma onto the caster and
drains them. (Your flesh rots into miasma!) You get one turn of miasma
immunity, but after that, if you're not undead, you'd better hustle out
of there. Cloud duration scales with power. (The drain is because this is
a very strong effect.)

Any creature which is vulnerable to miasma avoids it for dear life, so
it's a great way to make some space if you're in a jam. If you have some
way to trick monsters into it, though - Gell's Gravitas, Lesser Beckoning,
potions of attraction, orbs of mayhem... well, they're in for a bad time!
It's also hilariously good against Robin specifically.

The hope is that this will be more intuitive than Conjure Flame (with its
delayed 'embers' that monsters stomp on), and won't have the self-buff or
positioning issues of Corpse Rot. Also it hopefully won't be wildly
overpowered. Let's try it!

Replaces Conjure Flame in the Hedge Wizard start. Also available as a Kiku
gift.